backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
transformers
pillow
requests
numpy
scikit-learn
pydantic
